add exclusion list to antilock (avoid trigger when foreground window is in exclusion list) improve SerialForwarder improve GUI of config with bold titles small code optimizations set C++ version to C++20

+++ src/PathSeparator.cpp
@@ -15,7 +15,7 @@ void PathSeparator::ReplaceClipboard()
 			MyFrame* frame = ((MyFrame*)(wxGetApp().GetTopWindow()));
 			{
 				std::lock_guard lock(frame->mtx);
-				frame->pending_msgs.push_back({ (uint8_t)PathSeparatorsReplaced, std::move(input) });
+				frame->pending_msgs.push_back({ PathSeparatorsReplaced, std::move(input) });
 			}
 		}
 		wxTheClipboard->Close();
